requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
arrow==1.3.0
astrapy==1.1.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
binaryornot==0.4.4
bson==0.5.10
cassandra-driver==3.29.1
cassio==0.1.7
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cookiecutter==2.6.0
cookiecutter-data-science==2.0.0
dataclasses-json==0.6.5
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
geomet==0.2.1.post1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
keybert==0.8.5
langchain==0.1.17
langchain-astradb==0.2.0
langchain-community==0.0.37
langchain-core==0.2.34
langchain-openai==0.1.22
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.104
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.42.0
orjson==3.10.3
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pip-autoremove==0.10.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
pyzmq==26.1.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
sympy==1.13.2
tenacity==8.3.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.2
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uuid6==2024.1.12
wcwidth==0.2.13
yarl==1.9.4
